Adding two functions for walking packed protobufs
first, the function `protobuf_c_field_seek` is going to be a function
that walks packed protobuf data. This is adapted from
`protobuf_c_message_unpack` in protobuf-c. It will use a helper,
`parse_tag_and_wiretype`, to be able to find a specific field, and to
skip data for other fields.