Adding an initial structure for tests
This is a homegrown, very simple test layout that I am experimenting
with. The tests directory follows a similar patterns as toolbox, where
kctl is a tool in the toolbox directory, and getlog is a suite of tests
in the tests directory. Once some getlog tests are written, then a more
developed test framework may be adopted