del/get/put: updated extract functions
previously, the kv_keycnt variable was set to 1 or 0 depending on the
has_key attribute of the response. This is problematic, because even if
there is no key, the kiovec for the key is not cleared (by design).
Therefore, the extract function was updated to only change kv_keycnt if
has_key is true, but in that case it is set to 1 (aka, unchanged). But,
for generality the code is left in. This should be discussed or
documented in future design docs.