batch.c: updated extract_status
The extract_status() function was previously copy-pasted from an old
implementation for a command that uses kv_t. This version now uses
extract_cmdstatus defined in kinetic_protocol.h and implemented in
kinetic_protocol.c, which standardizes how to extract the status from
the response (and checks existence of data structures correctly).