batch.c: (#5) inverted clause and set krc
The logic for validating command sequences is many lines. To make the
other case readable--when sequences are not being validated--the
preprocessor condition was inverted (ifdef -> ifndef), krc is set
appropriately (the same as how it is set for start batch), and then the
code that was guarded by ifdef previously, is now in an else body.