protocol_interface.c: removed result vars
to address warnings for unused variables, hmac_result and repack_result,
I simply removed them. The "straightforward" way to use them was to not
pack/unpack the PDU so that it would have the wrong information.
However, if computing the hmac or repacking the message fail, then the
PDU will be incorrect anyways, so this has the same result. How we
handle the error will probably need to be decided in the future, though.