(#42) hashtable.[hc]pp: fixes and design changes
Most of the line differences will be fixes (e.g. names of kv_t
attributes). Some design changes are towards defining a TestKVEntry that
defines test-related conveniences that I don't expect are needed in the
general KVEntry definition. KVEntry itself is only in the test code, but
I'd like to possible promote to a general interface in the future.