fixtures/delete.bash: moved from valgrind-fixtures
I made the use of valgrind optional. When the test cases are called, if
they are provided "true" as the first argument, then valgrind will be
used. If the script is called directly, it will propagate this arg (2nd
arg; since 1st is kinetic IP). If sourced, then the argument must be
passed when the functions are called.

Also, I added implementations for the other test cases: atomic delete,
delete with cache policy, and delete range.