src/del.c: (#50) updated use of ki_addctx
previously, `ki_addctx` was invoked earlier, but because other failures
might happen afterward, there was an awkward nuance of having added to
the ktb_ctx for the delete keyval structure. By moving the call to
`ki_addctx` to be the last possible failure in `extract_delkey`, that
nuance disappears and `destroy_command` can be called without leaving a
dangling pointer.