getlog.c: (#50) fixes for getlog mem management
We previously would extract a getlog response into a new getlog
structure, and then memcpy the data and pointers into the provided
getlog structure. However, due to ktb memory management, the temporary
getlog structure was an **invalid** ktb type. The solution is to memcpy
pointers to a previous ktb ctx into a temporary stack-based structure.
On failure to extract a getlog response, the old pointers can be
recovered.